Added hittable and sphere modules
The hittable module defines the trait Hittable, which is implemented by
all structures which are hittable by a ray. In addition, it defines the
HittablesList struct, which is a collection of hittable structs.
Finally, it defines the HitRecord struct, which contains important
information relating to the ray-object collision.

The sphere module defines the Sphere struct, which implements Hittable.